Us 251 level objectives

+++ back-end/src/main/java/nl/hu/serious_game/domain/Objective.java
@@ -1,10 +1,10 @@
 package nl.hu.serious_game.domain;
 
 public class Objective {
-    private int maxCo2;
+    private float maxCo2;
     private int maxCoins;
 
-    public Objective(int maxCo2, int maxCoins) {
+    public Objective(float maxCo2, int maxCoins) {
         if (maxCo2 < 0 || maxCoins < 0) {
             throw new IllegalArgumentException("maxCo2 and maxCoins must be positive");
         }
@@ -12,7 +12,7 @@ public class Objective {
         this.maxCoins = maxCoins;
     }
 
-    public int getMaxCo2() {
+    public float getMaxCo2() {
         return maxCo2;
     }
